UPSTREAM: trogdor: Strappings_update_final3.1_second_thisistherealone.patch
Apparently what I thought was lazor-rev2 is actually lazor-rev3 and
nobody is really sure what lazor-rev4 is going to be at this point or
how we proceed from there. What seems to be somewhat agreed upon is that
for now all Lazor revisions use the "old" GPIO mapping and it's not very
clear if that's ever going to change for Lazor, so let's take the
revision restriction out from Lazor for now.

BUG=none
BRANCH=none
TEST=none

Change-Id: I29205555760c59f5f82141a66ca883d6985b3b57
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8678d47debc97bfe13c8e1bd2f0d99d8f8169b81
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I4939ccfd8464da6e72b5e01a58489b8c80f5b4df
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45225
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Douglas Anderson <dianders@chromium.org>
Original-Reviewed-by: Philip Chen <philipchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2409757
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>