UPSTREAM: mb/google/zork: Fix FPMCU_INT_L configuration
Fingerprint interrupt (FPMCU_INT_L) is level triggered and not edge
triggered. Also, we are using GEVENT for wake from fingerprint and
not the GPIO IRQ wake. Thus, the irq property exposed in ACPI tables
does not need to be set to indicate wake for the IRQ.

This change updates GPIO table to configure the pad as level triggered
and drops the wake attribute for irq_gpio in overridetree.

BUG=b:165612778
BRANCH=zork
TEST=Verified that fingerprint still works in S0 and to wake device
from S3.

Change-Id: I3aad605b043794ece2e5f86ac9c6bcdd02f3abf8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b0334e13babfa79ac6733d87c22fb93ef620a003
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I9007e5b0882ac1a6770db52d651218998f6d750d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45307
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2411976
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>