UPSTREAM: volteer: set GSPI CS to deasserted by default
This sets the state of GSPI chip select to 1 (deasserted) as applied
by the FSP during the silicon init phase. GSPI 0 and 1 are set to CS
mode manual in the SerialIoGSpiCsMode section which means we need to
explicitly configure CS to deasserted in the SerialIoGSpiCsState
section. GSPI0 is the CR50 and GSPI1 is the fingerprint sensor. We
were running into problems where the normal expected CS toggle
sequence to wake up CR50 did not work because CS was already asserted
when it was expected to be deasserted, leading to TPM timeouts.

BUG=b:168090038
TEST=booted on volteer, no more "TPM flow control failure" messages;
        verified fingerprint enrollment still works.

Change-Id: Ib13cb72fd041ba45d1c4a1c86d00c92359c32de3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 85e4c43b02da1561a6af589bdc21fdaa0f112777
Original-Change-Id: I47aa5db429d75e66095d58a1eb77963dcfc3b9f3
Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45384
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2421961
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>