UPSTREAM: drivers/soundwire/alc1308 : Add ALC1308 soundwire device
This patch adds new soundwire device ALC1308

The codec properties are filled out as best as possible
with the datasheet as a reference.

The ACPI address for the codec is calculated with the information in
the codec driver combined with the devicetree.cb hierarchy where the
link and unique IDs are extracted from the device path.

The unique ID is calculated from schematics by referring to ASEL[1:0]
strap settings. Datasheet of ALC1308 provides info about the mapping of
ASEL strap settings to unique ID

For example this device is connected to master link ID 1 and has strap
settings configuring it for unique ID 2.

chip drivers/soundwire/alc1308
  register "desc" = ""Left Speaker""
  device generic 1.2 on end
  end

Bug=None
Test=Build and boot on TGLRVP.Extract SSDT and confirm that the entries for
PCI0.HDAS.SNDW are present for ALC1308
Test speaker out functionality

BUG=none
BRANCH=none
TEST=none

Change-Id: I3df72d4272a90ebd27a079ce14d30cb6963d61c8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d7c31d1dbe89a54db8e244c2b0d746f58a1bd8ad
Original-Signed-off-by: Anil Kumar <anil.kumar.k@intel.com>
Original-Change-Id: Ibf3f1d5c6881cbd106e96ad1ff17ca216aa272ac
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51042
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-by: Sathyanarayana Nujella
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2726193
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>