UPSTREAM: sc7180: Remove the delay to force hpd detection and always disable HPD
HPD on this bridge chip is a bit useless. This is an eDP bridge so the HPD is
an internal signal that's only there to signal that the panel is done powering up.
But the bridge chip debounces this signal by between 100 ms and 400 ms (depending on process,
voltage, and temperate). One particular panel asserted HPD 84 ms after it was powered on
meaning that we saw HPD 284 ms after power on. Assume that the panel driver will have the
hardcoded delay in its prepare and always disable HPD.

BUG=none
BRANCH=none
TEST=none

Change-Id: I85ae638518e89db269f747a1dce803f6439773db
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b9a7d779b3cb48bb784d0856b6bb5f3fc6b4c1f7
Original-Change-Id: Iea7dd75b57fa55ec182c0bee09b0f35208357892
Original-Signed-off-by: Vinod Polimera <vpolimer@codeaurora.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45706
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2466032
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>