UPSTREAM: ppc64/bootblock_crt0.S: minimal implementation for bootblock C environment
BSS is loaded as part of the bootblock, it is zeroed in the file so it
doesn't have to be cleared explicitly by the code.

Code for clearing is left as a comment along with a warning about alignment
requirements.

Vector operations are sometimes generated for code such as
'uint8_t x[32] = {0}', this results in an exception when vector registers
(VR) are not enabled. VSR (vector-scalar register) operations are also
enabled, there is no reason not to.

(cherry picked from commit e527c713bd3533858054fe389ec2a5c93f6d6726)

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/57076
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Original-Change-Id: I878ef61619eb4a191805c8911d001312a0d717a0
GitOrigin-RevId: e527c713bd3533858054fe389ec2a5c93f6d6726
Change-Id: I24d8479ef37c44b48fbb4bbe41889f6c0f27ff50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3457013
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Andrew McRae <amcrae@google.com>