UPSTREAM: templates: add an empty SPD to SPD_SOURCES
Add an empty SPD in SPD_SOURCES when creating a new variant of
hatch, volteer, waddledee, or waddledoo, so that coreboot can build
successfully.

For variants that use spd_tools, add an empty mem_parts_used.txt so
that the developer can add the supported memory parts and regenerate
the Makefile.inc with the correct SPD references.

Add an empty SPD for LPDDR4x for waddledee and waddledoo to use.

BUG=b:169422833
TEST=create a new variant of hatch, volteer, waddledee, and waddledoo.
Observe that each one succeeds.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: f52e4a03ec30a497566c7860fbe790bfd9b9a166
Original-Signed-off-by: Paul Fagerburg <pfagerburg@google.com>
Original-Change-Id: I06dfb6103701bf8949180595f1e98fac48bcc585
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45746
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Change-Id: I07c3e780564051abe227ea056a4a406380f2f305
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2455143
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>