UPSTREAM: configs: Add a weird config for Portwell M107
This is not meant for actual use, but to build-test several options.
Please do not try to use it on real hardware. Or maybe do try.

The purpose of this config is to build-test the individual options, not
their combination. So, for instance, if it would be hard to keep options
x, y and z build together in the future, this config shouldn't block a
change but should instead be adapted, e.g. split into multiple chunks.

BUG=none
BRANCH=none
TEST=none

Change-Id: I50d6a399e74a910c4f8de3530caa80059964c880
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e8e0418d984f0ba4a6754ce37d6fe8ecc7cd4a2f
Original-Change-Id: Ife40d055e4c9b295c54cfc6a27af06e9358f7761
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45974
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2516469
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>