UPSTREAM: drivers/usb/acpi: Add support for privacy_gpio
Some devices, such as cameras, can implement a physical switch to
disable the input on demand. Think of it like the typical privacy
sticker on the notebooks, but more elegant.

In order to notify the system about the status this feature, a GPIO is
typically used.

The map between a GPIO and the feature is done via ACPI, the same way as
the reset_gpio works.

This patch implements an extra field for the described privacy gpio.
This gpio does not require any extra handling from the power management.

BUG=b:169840271

Change-Id: I44df9c4568317fc9d9af104fc12bb4572618f401
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: df79757a3f3d969b00fa198d50efcdd6517e00fe
Original-Change-Id: Idcc65c9a13eca6f076ac3c68aaa1bed3c481df3d
Original-Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46961
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2526552
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>