UPSTREAM: mrc_cache: Change mrc_cache_load_current to return size of entry
Modify mrc_cache_load current to return the size of the mrc_cache
entry so that caller will know what the actual size of the data
returned is.  This is needed for ARM devices like trogdor, which need
to know the size of the training data when populating the QcLib
interface table.

BUG=b:150502246
BRANCH=None
TEST=./util/abuild/abuild -p none -t GOOGLE_NAMI -x -a

Change-Id: Ic61b3c09ff35952c9fb31fe8109ebe0649ff4e2c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 93d483db89da3554afe01958e9f80a76251f2477
Original-Change-Id: Ia314717ad2a7d5232b37a19951c1aecd7f843c27
Original-Signed-off-by: Shelley Chen <shchen@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46110
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2466023
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>