UPSTREAM: soc/intel/tigerlake: Enable RTD3 driver and IPC mailbox
This SOC overrides the common PMC device and instantiates the PMC device
in the SSDT.  It needs to call the common PMC function to provide the
IPC mailbox method.

The common PCIe RTD3 driver can also be enabled which will allow
mainboards to enable Runtime D3 power control for PCIe devices.

BUG=b:160996445
TEST=boot on volteer with this driver enabled for the NVMe device in the
devicetree and disassemble the SSDT to ensure the RTD3 code is present.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: e997d85e3be50cda1ffdcf4d76014b713fe4951b
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: Ifa54ec3b8cebcc2752916cc4f8616fcb6fd2fecc
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46261
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: Ibb23f6dc1ee698dece5be01f8ab6515b5d29ab14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2553588
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>